+++ starter-dev-backend/package.json
@@ -8,13 +8,14 @@
   },
   "dependencies": {
     "@types/cookie-parser": "^1.4.2",
+    "apollo-server-lambda": "^3.11.1",
     "axios": "^0.24.0",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
     "express": "^4.17.1",
     "express-graphql": "^0.12.0",
-    "graphql": "^16.3.0",
+    "graphql": "^16.6.0",
     "morgan": "^1.10.0",
     "netlify-lambda": "^2.0.15",
     "serverless-http": "^2.7.0"
