--- starter-dev-backend/src/routes/index.js
@@ -1,7 +1,0 @@
-import Router from 'express';
-import graphql from './graphql';
-
-const routes = Router();
-routes.use('/', graphql);
-
-export default routes;
